LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/parsing/font-size-adjust-computed-expected.txt


PASS CSS Fonts Module Level 5: getComputedStyle().fontSizeAdjust
PASS Property font-size-adjust value 'none'
PASS Property font-size-adjust value '0.5'
PASS Property font-size-adjust value 'ex-height 0.5'
PASS Property font-size-adjust value 'cap-height 0.8'
PASS Property font-size-adjust value 'ch-width 0.4'
PASS Property font-size-adjust value 'ic-width 0.9'
PASS Property font-size-adjust value 'ic-height 1.1'
FAIL Property font-size-adjust value 'from-font' assert_equals: expected "0.5" but got "from-font"
FAIL Property font-size-adjust value 'ex-height from-font' assert_equals: expected "0.5" but got "from-font"
FAIL Property font-size-adjust value 'cap-height from-font' assert_equals: expected "cap-height 0.5" but got "cap-height from-font"
FAIL Property font-size-adjust value 'ch-width from-font' assert_equals: expected "ch-width 1" but got "ch-width from-font"
FAIL Property font-size-adjust value 'ic-width from-font' assert_equals: expected "ic-width 1" but got "ic-width from-font"
FAIL Property font-size-adjust value 'ic-height from-font' assert_equals: expected "ic-height 1" but got "ic-height from-font"
